services/core-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
requests==2.32.3
